backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
python-dotenv==1.0.0
pydantic==2.4.2
alembic==1.12.1
python-multipart==0.0.6
websockets==12.0